+++ CHANGELOG.md
@@ -84,6 +84,7 @@ Compatibility:
 * Implemented `rb_mProcess` (#1936).
 * Implemented `rb_gc_latest_gc_info` (#1937).
 * Implemented `RBASIC_CLASS` (#1935).
+* Yield 2 arguments for `Hash#map` if the arity of the block is > 1 (#1944).
 
 Changes: